# src/app/api/mcp/maze-game/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCPlayerOne Game API",
    "version": "1.0.0",
    "description": "API for interacting with the MCP text adventure game. Allows starting games, getting state, moving, taking items, examining things, and viewing leaderboards."
  },
  "servers": [
    {
      "url": "/api/game",
      "description": "Relative path to the game API endpoints"
    }
  ],
  "paths": {
...
    "/look": {
      "post": {
        "summary": "Look Around Current Location",
        "description": "Gets details about the player's current location (description, items, exits). Functionally similar to /state but may provide slightly different presentation.",
        "operationId": "lookAround",
        "tags": ["Game Actions"],
         "requestBody": {
          "description": "User and Story identifiers.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "location": { "$ref": "#/components/schemas/Location" },
                    "message": { "type": "string", "example": "You are in the Dragon Entrance..." },
                    "hint": { "type": "string", "example": "You see a Rusty Sword." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/examine": {
      "post": {
        "summary": "Examine Target",
        "description": "Gets a detailed description of a specific item or feature in the player's current location.",
        "operationId": "examineTarget",
        "tags": ["Game Actions"],
         "requestBody": {
          "description": "User, Story, and target Item/Feature identifiers.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamineInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Examination successful. Returns details about the target.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                   "properties": {
                    "success": { "type": "boolean", "example": true },
                    "name": { "type": "string", "example": "Rusty Sword" },
                    "description": { "type": "string", "example": "A standard-issue sword, showing signs of rust and neglect." },
                    "type": { "type": "string", "enum": ["item", "feature"], "example": "item" },
                    "hint": { "type": "string", "example": "It might be useful in a fight." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserStoryInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The unique identifier for the player.",
            "example": "user123"
          },
          "storyId": {
            "type": "string",
            "description": "The unique identifier for the story.",
            "example": "dragon_lair"
          }
        },
        "required": ["userId", "storyId"]
      },
...
      "ExamineInput": {
          "allOf": [
          { "$ref": "#/components/schemas/UserStoryInput" },
          {
            "type": "object",
            "properties": {
              "target": {
                "type": "string",
                "description": "The ID of the item or feature the player wants to examine.",
                "example": "rusty_sword"
              }
            },
             "required": ["target"]
          }
        ]
      },
...
      "Location": {
        "type": "object",
        "description": "Represents a location within a story. If an image is present, both a top-level 'image' and 'alt' field and a 'content' array are included. The content array contains a type: 'image' object first, followed by a type: 'text' object.",
        "properties": {
          "id": { "type": "string", "description": "Unique ID for the location within the story.", "example": "dragon_entrance" },
          "storyId": { "type": "string", "example": "dragon_lair" },
          "name": { "type": "string", "example": "Dragon Entrance" },
          "description": { "type": "string", "example": "You stand before a dark cave entrance..." },
          "items": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of item IDs currently present in the location.",
            "example": ["rusty_sword"] 
          },
           "features": { 
             "type": "array", 
             "items": { 
               "type": "object", 
               "properties": {
                 "id": {"type": "string", "example": "large_door"},
                 "name": {"type": "string", "example": "Large Door"},
                 "description": {"type": "string", "example": "A heavy oak door..."}
               }
             }, 
             "description": "List of examinable features in the location."
          },
          "exits": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of location IDs accessible from this location.",
            "example": ["guard_chamber"]
          },
           "requirements": {
             "type": "object",
             "properties": {
               "item": { "type": "string", "nullable": true, "description": "Item ID required to enter.", "example": "silver_key" },
               "condition": { "type": "string", "nullable": true, "description": "Condition ID (e.g., puzzle) required to enter.", "example": "gate_puzzle" }
             }
           },
          "image": { "type": "string", "description": "URL of the image for the location, if present.", "nullable": true },
          "alt": { "type": "string", "description": "Alt text for the image, if present.", "nullable": true },
          "content": {
            "type": "array",
            "items": {
              "oneOf": [
                { "$ref": "#/components/schemas/ImageContent" },
                { "$ref": "#/components/schemas/TextContent" }
              ]
            },
            "description": "Array with image and text content for the location."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "A message describing the error."
          },
          "hint": {
             "type": "string",
             "nullable": true,
             "description": "An optional hint for the user/agent."
          }
        },
        "required": ["success", "error"]
      },
...
      "ImageContent": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["image"] },
          "image": { "type": "string", "description": "URL of the image." },
          "alt": { "type": "string", "description": "Alt text for the image." }
        }
      },
      "TextContent": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["text"] },
          "text": { "type": "string", "description": "JSON stringified object with all metadata, including storyId." }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - The request was malformed or missing required parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found - The requested resource (player, story, location, item) could not be found.",
         "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error - Something went wrong on the server.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
} 